refactor(automanage): rename the detection task queue to automanage
Completes the detection→automanage rename started at the API layer (#471):
the queue itself was still `detection`. Straight rename, no behaviour change.

- Queue `detection` → `automanage` (queues.py, run_worker concurrency+metrics
  keys + module import, docker-compose worker-detection→worker-automanage, helm
  values + $metricsPorts).
- `app/tasks/detection.py` → `app/tasks/automanage.py`; review.py lazy import.
- Docstrings/tests updated; CLAUDE.md queue list corrected (was stale at four).
