+++ backend/tests/test_task_queue_metrics.py
@@ -101,7 +101,7 @@ def test_task_queue_collector_labels_every_queue(monkeypatch):
         "documents",
         "triggers",
         "coedit",
-        "detection",
+        "automanage",
         "lightweight_maintenance",
     }
     assert all(s.value == 2 for s in depth.samples)
